services/content: move service client into package
