+++ content/writer.go
@@ -102,7 +102,7 @@ func (w *writer) Commit(size int64, expected digest.Digest) error {
 	if err := os.Rename(ingest, target); err != nil {
 		if os.IsExist(err) {
 			// collision with the target file!
-			return errExists
+			return ErrExists
 		}
 		return err
 	}
